improve row number counting

+++ src/main/java/com/hagoapp/f2t/ParseObserver.java
@@ -39,6 +39,6 @@ public interface ParseObserver {
         return true;
     }
 
-    default void onRowCountDetermined(long rowCount) {
+    default void onRowCountDetermined(int rowCount) {
     }
 }
